Add abstract loader support (#220)
