+++ tests/SymfonyApp/TestBundle/DataFixtures/ORM/DataLoader.php
@@ -11,23 +11,6 @@
 
 namespace Hautelook\AliceBundle\Tests\SymfonyApp\TestBundle\DataFixtures\ORM;
 
-use Hautelook\AliceBundle\Doctrine\DataFixtures\AbstractLoader;
-
-class DataLoader extends AbstractLoader
+class DataLoader extends AbstractDataLoader
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function getFixtures()
-    {
-        \PHPUnit_Framework_Assert::assertInstanceOf(
-            'Symfony\Component\DependencyInjection\ContainerInterface',
-            $this->container
-        );
-
-        return [
-            __DIR__.'/product.yml',
-            __DIR__.'/brand.yml',
-        ];
-    }
 }
